+++ CHANGES.rst
@@ -4,7 +4,8 @@ Changelog
 0.13.14 (unreleased)
 --------------------
 
-- Nothing changed yet.
+- Let town decide if contact.core person properties fallback out of person's function or not. Add new registry params and upgrade steps.
+  [boulch]
 
 
 0.13.13 (2019-06-03)